2nd commit, added fetchAllArtists method, still having issues with createArtist due to values displaying null

+++ src/main/java/com/promineo/playlist/entity/Album.java
@@ -21,7 +21,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Album implements Comparable<Album> {
-  private Integer albumPk;
+  private Long albumPk;
   private String albumName;
   private Date releaseDate;
   private BigDecimal albumRating;
